feat(lld): move ui builder inside view instead of hook
